requirements.txt
# requirements.txt
groq
fastapi
uvicorn
python-dotenv
pydantic
requests
pandas
google-generativeai
google-generativeai>=0.7.0
